+++ demo/src/main/java/com/company/demo/services/PostService.java
@@ -7,14 +7,14 @@ import java.util.List;
 
 public interface PostService {
 
-    PostRequest createPost(PostRequest postRequest, Integer userId, Integer categoryId);
+    PostRequest createPost(PostRequest postRequest, Long userId, Integer categoryId);
     PostRequest updatePost(PostRequest postRequest, Integer postId);
     void deletePost(Integer postId);
     PostResponse getAllPost(Integer pageNumber, Integer pageSize, String sortBy);
     PostRequest getPostById(Integer postId);
 
     List<PostRequest> getAllPostByCategory(Integer categoryId);
-    List<PostRequest> getAllPostByUser(Integer userId);
+    List<PostRequest> getAllPostByUser(Long userId);
 
     //for searching:
     List<PostRequest> searchPosts(String keyword);
